push p errors

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: anmindt <anmindt@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/05 16:36:50 by anmindt           #+#    #+#             */
-/*   Updated: 2024/03/05 20:51:07 by anmindt          ###   ########.fr       */
+/*   Updated: 2024/03/08 19:20:25 by anmindt          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,7 @@
 # include <ctype.h>
 # include <stdio.h>
 # include <stdarg.h>
+# include <limits.h>
 
 //core functions for initialising printing with specific format
 int		ft_printf(const char *str, ...);
@@ -26,10 +27,10 @@ void	print_form(char str, int *count, va_list argptr);
 //functions for converting and printing numbers
 void	ft_print_nbr(int n, int *count);
 void	ft_print_unum(unsigned int n, int *count);
-int		ft_narrlen(unsigned long long n);
-char	*ft_narr(unsigned long long n);
+int		ft_narrlen(unsigned int n);
+char	*ft_narr(unsigned int n);
 void	ft_print_and_free(char *str, int *count);
-void	ft_dectohex_up(unsigned long long n, char c, int *count);
+void	ft_dectohex_up(unsigned int n, char c, int *count);
 
 //functions for printing chars and strings
 void	print_c(char c, int *count);
